+++ src/errors/unauthorized.error.ts
@@ -1,0 +1,7 @@
+import { AppError } from './app.error';
+
+export class UnauthorizedError extends AppError {
+  constructor(message = 'Unauthorized.') {
+    super(message, 'UnauthorizedError');
+  }
+}